refactor(products): colocate product catalog components
