+++ src/app/(main)/(products)/products/[category]/_containers/ProductCatalog.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useProducts, usePremiumFeature } from "@/ui/hooks";
-import { ProductCatalog as ProductCatalogView } from "@/ui/components/organisms";
+import { ProductCatalog as ProductCatalogView } from "../_components";
 import type { Product } from "@/services";
 import type { ProductCategory, SubCategory } from "@/core/domain";
