+++ src/app/(main)/(products)/products/[category]/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { ProductCatalog } from "./ProductCatalog";
