+++ src/ui/components/organisms/index.ts
@@ -7,9 +7,7 @@ export * from "./FormField";
 export * from "./ImageField";
 export * from "./ImagePreviewItem";
 export * from "./PersonValueCard";
-export * from "./ProductCatalog";
 export * from "./ProductFeatures";
-export * from "./ProductFilters";
 export * from "./ProductGrid";
 export * from "./ProductOptions";
 export * from "./ProductRegistrationForm";
